unify squad readers
