+++ allennlp_models/rc/dataset_readers/__init__.py
@@ -1,7 +1,6 @@
 from allennlp_models.rc.dataset_readers.drop import DropReader
 from allennlp_models.rc.dataset_readers.qangaroo import QangarooReader
 from allennlp_models.rc.dataset_readers.quac import QuACReader
-from allennlp_models.rc.dataset_readers.squad import SquadReader
-from allennlp_models.rc.dataset_readers.squad2 import Squad2Reader
+from allennlp_models.rc.dataset_readers.squad import BaseSquadReader, Squad1Reader, Squad2Reader
 from allennlp_models.rc.dataset_readers.transformer_squad import TransformerSquadReader
 from allennlp_models.rc.dataset_readers.triviaqa import TriviaQaReader
