+++ allennlp_models/rc/dataset_readers/squad.py
@@ -14,18 +14,25 @@ from allennlp_models.rc.dataset_readers import utils
 
 logger = logging.getLogger(__name__)
 
+SQUAD2_NO_ANSWER_TOKEN = "@@<NO_ANSWER>@@"
+"""
+The default 'no_answer_token' for the `Squad2Reader`.
+"""
 
-@DatasetReader.register("squad")
-class SquadReader(DatasetReader):
+
+class BaseSquadReader(DatasetReader):
     """
-    Reads a JSON-formatted SQuAD file and returns a ``Dataset`` where the ``Instances`` have four
-    fields: ``question``, a ``TextField``, ``passage``, another ``TextField``, and ``span_start``
-    and ``span_end``, both ``IndexFields`` into the ``passage`` ``TextField``.  We also add a
-    ``MetadataField`` that stores the instance's ID, the original passage text, gold answer strings,
-    and token offsets into the original passage, accessible as ``metadata['id']``,
-    ``metadata['original_passage']``, ``metadata['answer_texts']`` and
-    ``metadata['token_offsets']``.  This is so that we can more easily use the official SQuAD
-    evaluation script to get metrics.
+    Reads a JSON-formatted SQuAD 1.1/2.0 file and returns a `Dataset` where the `Instances` have the
+    following fields:
+
+      - `question`, a `TextField`,
+      - `passage`, another `TextField`,
+      - `span_start` and `span_end`, both `IndexFields` into the `passage` `TextField`,
+      - and `metadata`, a `MetadataField` that stores the instance's ID, the original passage text,
+        gold answer strings, and token offsets into the original passage, accessible as `metadata['id']`,
+        `metadata['original_passage']`, `metadata['answer_texts']` and
+        `metadata['token_offsets']`, respectively. This is so that we can more easily use the official
+        SQuAD evaluation scripts to get metrics.
 
     We also support limiting the maximum length for both passage and question. However, some gold
     answer spans may exceed the maximum passage length, which will cause error in making instances.
@@ -35,20 +42,24 @@ class SquadReader(DatasetReader):
     computed loss will not be accurate as a result. But this will not affect the answer evaluation,
     because we keep all the original gold answer texts.
 
-    # Parameters
-
+    Parameters
+    ----------
     tokenizer : `Tokenizer`, optional (default=`SpacyTokenizer()`)
-        We use this ``Tokenizer`` for both the question and the passage.  See :class:`Tokenizer`.
-        Default is `SpacyTokenizer()`.
+        We use this `Tokenizer` for both the question and the passage.  See :class:`Tokenizer`.
+        Default is ``SpacyTokenizer()`.
     token_indexers : `Dict[str, TokenIndexer]`, optional
         We similarly use this for both the question and the passage.  See :class:`TokenIndexer`.
         Default is `{"tokens": SingleIdTokenIndexer()}`.
-    passage_length_limit : `int`, optional (default=`None`)
-        if specified, we will cut the passage if the length of passage exceeds this limit.
-    question_length_limit : `int`, optional (default=`None`)
-        if specified, we will cut the question if the length of passage exceeds this limit.
-    skip_invalid_examples: `bool`, optional (default=`False`)
-        if this is true, we will skip those invalid examples
+    passage_length_limit : `int`, optional (default=None)
+        If specified, we will cut the passage if the length of passage exceeds this limit.
+    question_length_limit : `int`, optional (default=None)
+        If specified, we will cut the passage if the length of passage exceeds this limit.
+    skip_invalid_examples: `bool`, optional (default=False)
+        If this is true, we will skip invalid examples that don't contain the answer spans.
+    no_answer_token: `Optional[str]`, optional (default=`None`)
+        A special token to append to each context. If using a SQuAD 2.0-style dataset, this
+        should be set, otherwise an exception will be raised if an impossible question is
+        encountered.
     """
 
     def __init__(
@@ -58,7 +69,8 @@ class SquadReader(DatasetReader):
         passage_length_limit: int = None,
         question_length_limit: int = None,
         skip_invalid_examples: bool = False,
-        **kwargs
+        no_answer_token: Optional[str] = None,
+        **kwargs,
     ) -> None:
         super().__init__(**kwargs)
         self._tokenizer = tokenizer or SpacyTokenizer()
@@ -66,6 +78,7 @@ class SquadReader(DatasetReader):
         self.passage_length_limit = passage_length_limit
         self.question_length_limit = question_length_limit
         self.skip_invalid_examples = skip_invalid_examples
+        self.no_answer_token = no_answer_token
 
     @overrides
     def _read(self, file_path: str):
@@ -84,19 +97,28 @@ class SquadReader(DatasetReader):
 
                 for question_answer in paragraph_json["qas"]:
                     question_text = question_answer["question"].strip().replace("\n", "")
-                    answer_texts = [answer["text"] for answer in question_answer["answers"]]
-                    span_starts = [answer["answer_start"] for answer in question_answer["answers"]]
-                    span_ends = [
-                        start + len(answer) for start, answer in zip(span_starts, answer_texts)
-                    ]
+                    is_impossible = question_answer.get("is_impossible", False)
+                    if is_impossible:
+                        answer_texts: List[str] = []
+                        span_starts: List[int] = []
+                        span_ends: List[int] = []
+                    else:
+                        answer_texts = [answer["text"] for answer in question_answer["answers"]]
+                        span_starts = [
+                            answer["answer_start"] for answer in question_answer["answers"]
+                        ]
+                        span_ends = [
+                            start + len(answer) for start, answer in zip(span_starts, answer_texts)
+                        ]
                     additional_metadata = {"id": question_answer.get("id", None)}
                     instance = self.text_to_instance(
                         question_text,
                         paragraph,
-                        zip(span_starts, span_ends),
-                        answer_texts,
-                        tokenized_paragraph,
-                        additional_metadata,
+                        is_impossible=is_impossible,
+                        char_spans=zip(span_starts, span_ends),
+                        answer_texts=answer_texts,
+                        passage_tokens=tokenized_paragraph,
+                        additional_metadata=additional_metadata,
                     )
                     if instance is not None:
                         yield instance
@@ -106,45 +128,79 @@ class SquadReader(DatasetReader):
         self,  # type: ignore
         question_text: str,
         passage_text: str,
+        is_impossible: bool = None,
         char_spans: List[Tuple[int, int]] = None,
         answer_texts: List[str] = None,
         passage_tokens: List[Token] = None,
         additional_metadata: Dict[str, Any] = None,
     ) -> Optional[Instance]:
-
         if not passage_tokens:
             passage_tokens = self._tokenizer.tokenize(passage_text)
-        question_tokens = self._tokenizer.tokenize(question_text)
-        if self.passage_length_limit is not None:
+
+        if self.no_answer_token is not None:
+            if self.passage_length_limit is not None:
+                passage_tokens = passage_tokens[: self.passage_length_limit - 1]
+            passage_tokens = passage_tokens + [
+                Token(
+                    text=self.no_answer_token,
+                    idx=passage_tokens[-1].idx + len(passage_tokens[-1].text) + 1,  # type: ignore
+                    lemma_=self.no_answer_token,
+                )
+            ]
+        elif self.passage_length_limit is not None:
             passage_tokens = passage_tokens[: self.passage_length_limit]
+
+        question_tokens = self._tokenizer.tokenize(question_text)
         if self.question_length_limit is not None:
             question_tokens = question_tokens[: self.question_length_limit]
-        char_spans = char_spans or []
-        # We need to convert character indices in `passage_text` to token indices in
-        # `passage_tokens`, as the latter is what we'll actually use for supervision.
-        token_spans: List[Tuple[int, int]] = []
-        passage_offsets = [(token.idx, token.idx + len(token.text)) for token in passage_tokens]
-        for char_span_start, char_span_end in char_spans:
-            if char_span_end > passage_offsets[-1][1]:
-                continue
-            (span_start, span_end), error = utils.char_span_to_token_span(
-                passage_offsets, (char_span_start, char_span_end)
-            )
-            if error:
-                logger.debug("Passage: %s", passage_text)
-                logger.debug("Passage tokens: %s", passage_tokens)
-                logger.debug("Question text: %s", question_text)
-                logger.debug("Answer span: (%d, %d)", char_span_start, char_span_end)
-                logger.debug("Token span: (%d, %d)", span_start, span_end)
-                logger.debug("Tokens in answer: %s", passage_tokens[span_start : span_end + 1])
-                logger.debug("Answer: %s", passage_text[char_span_start:char_span_end])
-            token_spans.append((span_start, span_end))
-        # The original answer is filtered out
-        if char_spans and not token_spans:
-            if self.skip_invalid_examples:
-                return None
-            else:
-                token_spans.append((len(passage_tokens) - 1, len(passage_tokens) - 1))
+
+        if is_impossible is True:
+            if self.no_answer_token is None:
+                raise ValueError(
+                    "This is a SQuAD 2.0 dataset, yet your using a SQuAD reader has 'no_answer_token' "
+                    "set to `None`. "
+                    "Consider specifying the 'no_answer_token' or using the 'squad2' reader instead, which "
+                    f"by default uses '{SQUAD2_NO_ANSWER_TOKEN}' as the 'no_answer_token'."
+                )
+            answer_texts = [self.no_answer_token]
+            token_spans: List[Tuple[int, int]] = [
+                (len(passage_tokens) - 1, len(passage_tokens) - 1)
+            ]
+        else:
+            char_spans = char_spans or []
+            # We need to convert character indices in `passage_text` to token indices in
+            # `passage_tokens`, as the latter is what we'll actually use for supervision.
+            token_spans = []
+            passage_offsets = [(token.idx, token.idx + len(token.text)) for token in passage_tokens]
+            for char_span_start, char_span_end in char_spans:
+                if char_span_end > passage_offsets[-1][1]:
+                    continue
+                (span_start, span_end), error = utils.char_span_to_token_span(
+                    passage_offsets, (char_span_start, char_span_end)
+                )
+                if error:
+                    logger.debug("Passage: %s", passage_text)
+                    logger.debug("Passage tokens (with no-answer): %s", passage_tokens)
+                    logger.debug("Question text: %s", question_text)
+                    logger.debug("Answer span: (%d, %d)", char_span_start, char_span_end)
+                    logger.debug("Token span: (%d, %d)", span_start, span_end)
+                    logger.debug(
+                        "Tokens in answer: %s",
+                        passage_tokens[span_start : span_end + 1],
+                    )
+                    logger.debug("Answer: %s", passage_text[char_span_start:char_span_end])
+                token_spans.append((span_start, span_end))
+            # The original answer is filtered out
+            if char_spans and not token_spans:
+                if self.skip_invalid_examples:
+                    return None
+                else:
+                    token_spans.append(
+                        (
+                            len(passage_tokens) - 1,
+                            len(passage_tokens) - 1,
+                        )
+                    )
         return utils.make_reading_comprehension_instance(
             question_tokens,
             passage_tokens,
@@ -153,4 +209,65 @@ class SquadReader(DatasetReader):
             token_spans,
             answer_texts,
             additional_metadata,
+        )
+
+
+@DatasetReader.register("squad")
+class Squad1Reader(BaseSquadReader):
+    """
+    Dataset reader for SQuAD v1.1-style datasets. Arguments are the same as `BaseSquadReader`.
+
+    Registered as 'squad'.
+    """
+
+    def __init__(
+        self,
+        tokenizer: Tokenizer = None,
+        token_indexers: Dict[str, TokenIndexer] = None,
+        passage_length_limit: int = None,
+        question_length_limit: int = None,
+        skip_invalid_examples: bool = False,
+        **kwargs,
+    ) -> None:
+        super().__init__(
+            tokenizer=tokenizer,
+            token_indexers=token_indexers,
+            passage_length_limit=passage_length_limit,
+            question_length_limit=question_length_limit,
+            skip_invalid_examples=skip_invalid_examples,
+            no_answer_token=None,
+            **kwargs,
+        )
+
+
+@DatasetReader.register("squad2")
+class Squad2Reader(BaseSquadReader):
+    """
+    Dataset reader for SQuAD v2.0-style datasets. Arguments are the same as `BaseSquadReader`.
+
+    The only difference between this and `Squad1Reader` is that this reader appends a
+    special 'no_answer_token' to the passage field. For questions that are marked impossible,
+    the correct answer will just be the 'no_answer_token'.
+
+    Registered as 'squad2'.
+    """
+
+    def __init__(
+        self,
+        tokenizer: Tokenizer = None,
+        token_indexers: Dict[str, TokenIndexer] = None,
+        passage_length_limit: int = None,
+        question_length_limit: int = None,
+        skip_invalid_examples: bool = False,
+        no_answer_token: str = SQUAD2_NO_ANSWER_TOKEN,
+        **kwargs,
+    ) -> None:
+        super().__init__(
+            tokenizer=tokenizer,
+            token_indexers=token_indexers,
+            passage_length_limit=passage_length_limit,
+            question_length_limit=question_length_limit,
+            skip_invalid_examples=skip_invalid_examples,
+            no_answer_token=no_answer_token,
+            **kwargs,
         )
